+++ src/pulumi/core-dtap/analytics/databricks/workspaces/engineering.py
@@ -170,10 +170,7 @@ datafactory_token = databricks.Token(
 # ----------------------------------------------------------------------------------------------------------------------
 
 # If no clusters are defined in the YAML files, we'll not attempt to create any.
-try:
-    cluster_definitions = workspace_config["clusters"]
-except:
-    cluster_definitions = {}
+cluster_definitions = workspace_config.get("clusters", {})
 
 clusters = {}
 for ref_key, cluster_config in cluster_definitions.items():
@@ -184,17 +181,16 @@ for ref_key, cluster_config in cluster_definitions.items():
     )
 
     # Cluster Libraries
-    cluster_libraries = []
-    try:
-        # PyPi
-        for lib in cluster_config["libraries"]["pypi"]:
-            cluster_libraries.append(
-                databricks.ClusterLibraryArgs(
-                    pypi=databricks.ClusterLibraryPypiArgs(package=lib["package"])
-                )
+    pypi_libraries = [
+        databricks.ClusterLibraryArgs(
+            pypi=databricks.ClusterLibraryPypiArgs(
+                package=lib.get("package"), repo=lib.get("repo")
             )
-    except KeyError:
-        pass
+        )
+        for lib in cluster_config.get("libraries", {}).get("pypi", [])
+    ]
+
+    cluster_libraries = [*pypi_libraries]
 
     base_cluster_configuration = {
         "resource_name": cluster_resource_name,