+++ src/pulumi/core-dtap/analytics/datafactory/orchestration.py
@@ -1,13 +1,20 @@
+from pulumi import ResourceOptions
 from pulumi_azure_native import datafactory as adf, keyvault
 
-from ingenii_azure_data_platform.iam import GroupRoleAssignment, \
-    ServicePrincipalRoleAssignment, UserAssignedIdentityRoleAssignment
+from ingenii_azure_data_platform.iam import (
+    GroupRoleAssignment,
+    ServicePrincipalRoleAssignment,
+    UserAssignedIdentityRoleAssignment,
+)
 from ingenii_azure_data_platform.orchestration import AdfSelfHostedIntegrationRuntime
 from ingenii_azure_data_platform.utils import generate_resource_name
 
 from analytics.databricks.workspaces import engineering as databricks_engineering
-from platform_shared import get_devops_principal_id, get_devops_config_registry, \
-    get_devops_config_registry_resource_group
+from platform_shared import (
+    get_devops_principal_id,
+    get_devops_config_registry,
+    get_devops_config_registry_resource_group,
+)
 from project_config import platform_config, platform_outputs
 from management import resource_groups
 from management.user_groups import user_groups
@@ -247,7 +254,7 @@ databricks_file_ingestion_trigger = adf.Trigger(
 UserAssignedIdentityRoleAssignment(
     role_name="Data Factory Contributor",
     principal_id=get_devops_principal_id(),
-    scope=datafactory.id
+    scope=datafactory.id,
 )
 
 keyvault.Secret(
@@ -255,7 +262,5 @@ keyvault.Secret(
     resource_group_name=get_devops_config_registry_resource_group(),
     vault_name=get_devops_config_registry()["key_vault_name"],
     secret_name=f"data-factory-name-{platform_config.stack}",
-    properties=keyvault.SecretPropertiesArgs(
-        value=datafactory.name
-    ),
+    properties=keyvault.SecretPropertiesArgs(value=datafactory.name),
 )